# internal/docs/swagger.yaml
basePath: /v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  core.PlateAppearanceLeverage:
    properties:
      away_score_before:
        type: integer
      bases_before:
        description: '"100", "011", etc.'
        type: string
      batter_id:
        type: string
      description:
        description: Result summary.
        type: string
      event_id:
        description: index within game
        type: integer
      game_id:
        type: string
      home_score_before:
        type: integer
      inning:
        type: integer
      li:
        description: LI at start of PA
        type: number
      outs_before:
        type: integer
      pitcher_id:
        type: string
      top_of_inning:
        type: boolean
      we_after:
        type: number
      we_before:
        description: home team WE
        type: number
      we_change:
        description: signed WEAfter - WEBefore
        type: number
    type: object
...
info:
  contact:
    email: info@stormlightlabs.org
    name: API Support
    url: https://github.com/stormlightlabs/baseball
  description: |
    <!-- markdownlint-disable -->

    A comprehensive REST API for baseball statistics serving data from the Lahman Baseball Database and Retrosheet

    ## Attributions

    **[SABR Lahman Database](https://sabr.org/lahman-database/)**: The information used here was obtained free of charge from and is copyrighted by Sean Lahman.

    **[Retrosheet.org](https://www.retrosheet.org/)**: The information used here was obtained free of charge from and is copyrighted by Retrosheet.

    wOBA weights and seasonal constants from **[FanGraphs Guts!](https://www.fangraphs.com/tools/guts)**, used under FanGraphs’ published terms.
  license:
    name: MPL-2.0
    url: https://opensource.org/license/mpl-2-0
  title: Baseball API
  version: "1.0"
paths:
...
  /games/{game_id}/plate-appearances/leverage:
    get:
      consumes:
      - application/json
      description: Get leverage index for each plate appearance in a game
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      - default: 0
        description: Minimum leverage index
        in: query
        name: min_li
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.PlateAppearanceLeverage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get plate appearance leverages
      tags:
      - computed
      - games
      - leverage
...
  /games/{game_id}/win-probability/summary:
    get:
      consumes:
      - application/json
      description: Get summary statistics for a game's win probability including biggest
        swings
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.GameWinProbabilitySummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get game win probability summary
      tags:
      - computed
      - games
      - leverage
...
  /parks/{park_id}/factors:
    get:
      consumes:
      - application/json
      description: Get park factor for a specific park and season
      parameters:
      - description: Park ID
        in: path
        name: park_id
        required: true
        type: string
      - description: Season year
        in: query
        name: season
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ParkFactor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get park factor
      tags:
      - computed
      - parks
  /parks/{park_id}/factors/series:
    get:
      consumes:
      - application/json
      description: Get park factors over a range of seasons
      parameters:
      - description: Park ID
        in: path
        name: park_id
        required: true
        type: string
      - description: Starting season
        in: query
        name: from_season
        required: true
        type: integer
      - description: Ending season
        in: query
        name: to_season
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.ParkFactor'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get park factor series
      tags:
      - computed
      - parks
...
  /players/{player_id}/leverage/high:
    get:
      consumes:
      - application/json
      description: Get high-leverage plate appearances for a player in a season
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - default: 1.5
        description: Minimum leverage index
        in: query
        name: min_li
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.PlateAppearanceLeverage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get high leverage plate appearances
      tags:
      - computed
      - players
      - leverage
  /players/{player_id}/leverage/summary:
    get:
      consumes:
      - application/json
      description: Get aggregated leverage metrics for a player in a season
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - default: '""'
        description: Role filter (batter, pitcher)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.PlayerLeverageSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get player leverage summary
      tags:
      - computed
      - players
      - leverage
...
  /players/{player_id}/stats/baserunning:
    get:
      consumes:
      - application/json
      description: Get baserunning value (wSB) for a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.BaserunningStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get baserunning stats
      tags:
      - computed
      - players
  /players/{player_id}/stats/batting/advanced:
    get:
      consumes:
      - application/json
      description: Get wOBA, wRC+, ISO, BABIP, and other advanced batting stats for
        a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.AdvancedBattingStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get advanced batting stats
      tags:
      - computed
      - players
      - batting
  /players/{player_id}/stats/fielding:
    get:
      consumes:
      - application/json
      description: Get fielding runs above average for a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.FieldingStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get fielding stats
      tags:
      - computed
      - players
  /players/{player_id}/stats/pitching/advanced:
    get:
      consumes:
      - application/json
      description: Get FIP, xFIP, ERA+, and other advanced pitching stats for a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.AdvancedPitchingStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get advanced pitching stats
      tags:
      - computed
      - players
      - pitching
  /players/{player_id}/stats/war:
    get:
      consumes:
      - application/json
      description: Get Wins Above Replacement and component breakdown for a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 2024
        description: Season year
        in: query
        name: season
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.PlayerWARSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get WAR
      tags:
      - computed
      - players
...
swagger: "2.0"
...